Bring over HDF5 warning macros + cleanup (#713)
* Brings over H5_GCC_CLANG_DIAG_OFF, etc. macros
* Clean up more warnings
* Suppress a few unavoidable warnings, mainly functions taking
  const pointers via in/out pointers when the context is "write"
  and the buffer will be unmodified. We can't fix this without
  public API changes.

+++ hdf/src/dfconv.c
@@ -400,7 +400,7 @@ DFconvert(uint8 *source, uint8 *dest, int ntype, int sourcetype, int desttype, i
     }
 
     if (sourcetype == desttype) {
-        memcpy(dest, source, size);
+        memcpy(dest, source, (size_t)size);
         return 0;
     }
